Adds date to saved canvas filename
This change makes canvas saves ('P' key) save to a filename with date and time, rather than just 'canvas.png'.  This prevents issues with browser auto-renaming (Chrome calls it quits after 100, for example), as well as file management issues for the user (if they saved 1-10, and deleted 5, the new file would be saved as 5 again instead of continuing the count at 11).  The format saved to is: "pxls canvas YYYY-MM-DD HH.MM.SS.png" and *should* be independent of locale.