Ability to lock the canvas
When placing multiple pixels very quickly you often encounter an issue were the you start dragging around instead of placing pixels. When working on big drawings this is specially annoying since instead of placing you end up moving the canvas away. 

This PR proposes a way to prevent this by pressing L to toggle the locked state and preventing the user to drag the canvas around or zoom into it.