Fix Chrome template rendering at 100% scale
Chrome's nearest neighbor (pixelated) bitmap rendering is wonk when the image is offset by fractional pixels.  For dotted (3x3) templates, this means that they seem to disappear from view.  Using integer pixel offsets fixes this.  Not an issue in FireFox, but fix doesn't hurt it either.  As fractional pixel offsets are meaningless at scales of 1 and below (zooming out), the offsets are rounded at these scales.
Note that I chose to round the value in the code that applies the CSS, rather than rounding the internal value, as I haven't explored any (minor) ramifications that would bring when scaling back up (zooming in).