Resolving conflict with evt.code and jQuery version used
This change should resolve a conflict with the use of <keyboardEvent>.code and the version of the jQuery library used; the jQuery-modified event object does not include .code .  This causes that part of the switch statement to always be skipped as it reads 'undefined'.  As a result, some log-standing bugs - like the Mute key on some keyboards with some browsers causing the canvas to zoom out - still rear their head.

This change causes .code to be read from the original event object instead.  If not available (old browser or platform) it will still skip to the next test.
This change was not tested within a pxls config context, so please review carefully and make sure that all the .code events for possible keys do match up? :x