Adjust scrollIntoView() params for iframes
On FireFox and possibly Safari (Edge:Unknown)
This change adds the block/inline 'nearest' parameters to the palette selection scrollIntoView() method.  Previously, without specifying, the 'block' parameter would default to "start". 'block' defines how the element being scrolled into view is scrolled vertically.  While on pxls.space proper this is a moot issue as none of the parent containers have vertical scrolling, an issue presents itself if pxls.space is placed within an iframe in a larger document, such as in the pxlsFiddle Avogadro setup, where some browsers - FireFox confirmed - will cause the parent document to match the scroll behavior and thus would place the palette at the top of the user's screen.  With the palette at the bottom of the pxls UI, this means the canvas would no longer be visible.  
The reason "nearest" is chosen, rather than "end" (which would effectively position the palette at the bottom of the parent document) is that this behavior will leave the vertical scroll be if the element is already visible within the vertical client area, which should provide the east impact to the parent document.

This change should affect *only* the aforementioned browser and specifically described scenario, and not behavior under different browsers or with pxls.space as the top level document.