Added vibration option
Added code that makes a device vibrate for the amount of time, in ms, specified through the user interface and stored in localStorage.

Vibration is a resource (battery) hungry effect, and thus defaults to 0ms.  The code checks two things before deciding to vibrate: the browser must suggest that it supports vibrate (note: this does not guarantee that it *can* vibrate, and most browsers will happily suggest that they can as part of anti-fingerprinting measures), and whether or not the result off ls.get() is valid and non-zero (via simple validity test).

Because vibration is set up as part of the alert, some of the internal functions and calls have been renamed from Audio-specific, to Alert-specific.