Connection Idling
Idle out connections that haven't placed in a while. This should help
the cooldown not balloon when people leave their computers on at night.

A user is counted as "idle" when their lastPixelTime is less than the
configured amount. Once they place a pixel they're immediately counted
as active again and the user count updates to reflect it.