Add color to #cursor-text rather than using the default color
Now that color is properly set in the body, this text which relied on
the default being black became obscured when using themes that changed
the base text color to white. Adding a fixed color to contrast the
background fixes this.

Another solution would have been to add variables to this element but I
have decided against that for the same reason I initially did: the
existing themes chose not to style it.