Use standard classes for admin panel and lookup
The admin box located on the center-left of the screen uses similar
theming to other UI elements of the class `bubble`. Similarly, the mod
user lookup is styled as the pixel lookup UI.

This commit codifies these similarities by stripping some styling out
of admin-specific rules, applying the `bubble` class to the sidepanel,
and creating a new class `floating panel` which is applied to both the
pixel lookup and the user lookup UI components.

These changes allow us to collapse the admin CSS variables into
existing ones, making theming simpler and more coheasive.