Implement CSS Variables (#318)
* Add css variables to primary stylesheets

Since css allows fallbacks, browsers that do not support css variables
will use the original definitions and browsers that do have support will
use the new css variable ones.

While the original styles should be mostly untouched, this does
occasionally add definitions where none previously existed. Most of
these are color assignments which are simply set to inherit - the
default color value - resulting in no change. The most notable exception
is in inputs (textarea, select, button, etc.) which now set color,
background-color, and border.

* Update themes to use css variables

Besides a few edge-cases, this updates themes to work entirely on
setting css variables.

The edge cases are as follows:
- hr definitions have been left.
- dark.css seems to have an incorrect rule which has been left alone.
- darker.css defines some new borders.
   This is notable because there is no simple elegant solution to this
   with css variables. While it may be resolved in future, I have left
   it for now.
- Keyframe definitions for -scrolled-flash have been left static.
   In testing css variables appeared to not function when used to set
   the animation properties. This might be a bug in Firefox or something
   else resolvable but I'm not willing to do that at this time.

* Remove unused rules

* Add color to #cursor-text rather than using the default color

Now that color is properly set in the body, this text which relied on
the default being black became obscured when using themes that changed
the base text color to white. Adding a fixed color to contrast the
background fixes this.

Another solution would have been to add variables to this element but I
have decided against that for the same reason I initially did: the
existing themes chose not to style it.

* Remove button elements from generic input styling

* Fix dark theme lookup text being the wrong color

* Remove and standardise dev comments