Format modal dialogs like panels
This does a pretty rough job of creating modals using the same structure
as the side panels so that both may use the same styling.

One of the major compromises that I'm not happy about with this is the
requirement that the default close button not be shown. Since this is
controllable (or even forcible in modal creation), it shouldn't pose too
much issue but it still feels concerning to, along with the other
changes, modify and strip out so much of the original modal
functionality and structure.

This commit leaves things in a pretty rough state but future commits
should seek to clean things up and bring things in closer alignment.