Make controls button elements
This includes the main buttons for opening the panels, the icon buttons
for closing them, the buttons for opening popups, and the palette
selection buttons.

This also creates a new class `text-button` which serves the same role
as the previously removed `button` class. buttons are now mostly
unstyled by default, requiring them to be given some explit styling
before apearing functional.