Restore panel widths to 30%
The original code was a mess, resulting in them always taking as little
as possible which was the min-width of 30%. Now that things have been
cleaned up a bit, they take as much as needed up to the maximum. Since
they often include long sentences or paragraphs, this is usually
max-width which was 50% but is now 30% to maintain consistency.