Fix halfwidth panels and ui adaptation to panels
Previously the main ui attempted to resize itself to stay visible when a
panel was on screen. This commit makes this more effective by fixing
various bugs related to halfwidth and horizontal panels as well as
fixing the size of each panel so that the main ui can more effectively
resize itself.

This commit also reintroduces half-width panels and makes the help panel
such a panel as it was before.