Replace scale adjust algorithm and add configurable smooth zoom (#325)
In order to support smooth scrolling from devices capable of
pixel-perfect scrolling, the nudgeScale function needed to be able to
accept any value. This lead to its reimplementation as something akin to
what it originally was - a simple exponential function. This type of
function works well for scaling as percieved changes in size per fixed
unit scaled should be the same at any zoom level.

To mitigate the concern that any reasonable default is likely to be seen
as too sensitive by a handful of people, this also implements a setting
to change the base of this function from anywhere between 1.01 and 3
with the default being 1.5.