CSS cleanup (#324)
* Remove unused `darker` class

There are no longer any CSS rules for this class - it appears to be
old code that was never properly cleaned.

* Unify buttons

Buttons before were represented in one of three ways:
- A `<button>` by itself
- A `<button>` with the `button` class
- Another element with the `button` class

This removes the latter two representations and makes all styles that
would have applied to the `button` class instead apply to the button
element.

As a footnote, this commit also removes rules for buttons with the
unused `dark` css class.

* Use standard classes for admin panel and lookup

The admin box located on the center-left of the screen uses similar
theming to other UI elements of the class `bubble`. Similarly, the mod
user lookup is styled as the pixel lookup UI.

This commit codifies these similarities by stripping some styling out
of admin-specific rules, applying the `bubble` class to the sidepanel,
and creating a new class `floating panel` which is applied to both the
pixel lookup and the user lookup UI components.

These changes allow us to collapse the admin CSS variables into
existing ones, making theming simpler and more coheasive.

* Use `.floating-panel` for `.message` popups

This also applies the box-shadow of message popups to the floating
panel class and cuts down on a couple more CSS variables.

This also separates the rules pertaining to the `content` class. This
class was being used both for chat content and for floating-panel
(lookup) data.

* Remove `buton` class

The rest of these were already removed - this one was not picked up
during that due to the mistaken spelling of `button`.

* Format modal dialogs like panels

This does a pretty rough job of creating modals using the same structure
as the side panels so that both may use the same styling.

One of the major compromises that I'm not happy about with this is the
requirement that the default close button not be shown. Since this is
controllable (or even forcible in modal creation), it shouldn't pose too
much issue but it still feels concerning to, along with the other
changes, modify and strip out so much of the original modal
functionality and structure.

This commit leaves things in a pretty rough state but future commits
should seek to clean things up and bring things in closer alignment.

* Isolate header rules

Currently all `<header>`s are presumed to be used in a similar way,
requiring a lot of bending-over backwards to get a regular semantic
header with no styling of a particular kind.

This commit should resolve this, transferring a lot of the rules to
apply specifically to panel headers.

* Make controls button elements

This includes the main buttons for opening the panels, the icon buttons
for closing them, the buttons for opening popups, and the palette
selection buttons.

This also creates a new class `text-button` which serves the same role
as the previously removed `button` class. buttons are now mostly
unstyled by default, requiring them to be given some explit styling
before apearing functional.

* Style notifications like panel articles

* Use static-flow for .panel-body instead of absolute positioning

* Style chatban history like panel articles

* Use button element for show emoji button

* Use button element for reset typeahead button

* Use preformatted whitespace when displaying message in popup

* Use standard text buttons for chat popup menu

* Style chat popup like panels

* Use button element for chat jump to bottom button

Also adds some CSS variables for that button.

* Use .message instead of specific code for admin lookup

* Tidy up styling

This includes things such as removing unused or unnecessary code,
fixing minor bugs introduced with recent changes, and moving rules
into more appropriate sections.

* Remove outdated inline

These rules no longer select anything and I haven't heard any
complaints.

* Rename report-button to dangerous-button

* Fix layering of mod lookup

* Use dangerous-button in more places

* Move buttons in pixel report out of footer

* Use button elements for typeahead options

* Restore panel widths to 30%

The original code was a mess, resulting in them always taking as little
as possible which was the min-width of 30%. Now that things have been
cleaned up a bit, they take as much as needed up to the maximum. Since
they often include long sentences or paragraphs, this is usually
max-width which was 50% but is now 30% to maintain consistency.

* Use footer text color for article footers instead of header text color

* Update purple theme to use new variables

* Replace background-color variables with just background

This gives the flexibility to themes to add things like gradients and
images to backgrounds. This has the added benefit of also making
background variable naming more consistent.

* Compact .floating-panel and .message to be more consistent

* Renamed and fixed incorrect --emoji-picker-emoji-search-* variables

* Fix horizontal chat not working

The old CSS rules that handled this were previously removed because it
was thought they did nothing. As it turns out, they did something.

* Fix halfwidth panels and ui adaptation to panels

Previously the main ui attempted to resize itself to stay visible when a
panel was on screen. This commit makes this more effective by fixing
various bugs related to halfwidth and horizontal panels as well as
fixing the size of each panel so that the main ui can more effectively
resize itself.

This commit also reintroduces half-width panels and makes the help panel
such a panel as it was before.

* Fix empty ping list indicator being clipped

* Remove unused classes from faq and info

Co-authored-by: Andrew Black <andrewblack2656@gmail.com>