Fix chrome enabling multiple themes at once
Since Chrome does not fire the load event for elements that have already
loaded being added back to the DOM, the code that cleaned up previous
themes would not fire when switching back to a theme that had previously
been loaded.

This fix changes the way themes are loaded such that they stay on the
page and are disabled with the `disabled` property until needed again.
In addition, a `loaded` variable is also stored to determine if the app
needs to add the newly selected theme to the DOM.