Fix aside panel headings being one higher level than expected
The title of each aside panel is heading level 2. This means that inside
each panel we expect every heading to be at level 3 or below, so as to
stay under that same level 2 heading for the entire panel. This was not
the case as panel contents used level 2 headings for subsections. This
commit resolves this by bumping these level 2 headings to level 3 and
bumping any nested headings appropriately.