Use a Service Worker when possible to detect which tab has focus
The Service Worker functions as a middle man: recieving "focus" messages a tab (window) and broadcasting it to all other tabs.
It will also handle "request-id" messages, which will send back to the window their internal client ID which is then set in `uiHelper.tabId`
An improvement over the localStorage is that when a window is closed, it will send a "leave" message to the worker and the worker will designate focus to another window.

This commit also exposes `uiHelper.tabId` and `uiHelper.tabHasFocus()` to the App namespace.