Fix info links layout and make minor semantic changes (#344)
* Wrap info links in a list structure

This returns the previous formatting of being on one line each. The last
two links were put in a separate list as this seemed like the most
approriate meaning intended by the break originally placed there.

* Fix aside panel headings being one higher level than expected

The title of each aside panel is heading level 2. This means that inside
each panel we expect every heading to be at level 3 or below, so as to
stay under that same level 2 heading for the entire panel. This was not
the case as panel contents used level 2 headings for subsections. This
commit resolves this by bumping these level 2 headings to level 3 and
bumping any nested headings appropriately.

* Group board and its ui with `main` tag

While this should have little to no functional effect, it does make an
improvement - however small - to the semantics of the document.

* Use h2 for admin sidepanel instead of h1

* Add padding to info links