Clamp final rendering scale to integer values
This prevents a number of rendering artifacts present when scaling is
non-integer. This includes different columns and rows of pixels being
different sizes at the same scale and parts of the grid being cut-off.

The downside to this approach is that scrolling is now a lot less
smooth - something most readilly noticable when using a pixel-perfect
scrolling device such as a trackpad.