Refactor settings (#337)
* Create settings namespace and use it to manage settings

Chat settings have been excluded for now. This exposes the ability to
get, set, reset, listen for changes, disable, and enable most of the
settings found on the options panel. Some notable exceptions are
template options and Discord account settings since both of these are
handled without the use of localstorage.

This commit also leaves the original names for all associated settings
intact but, given how horribly inconsistent they are, I expect a
future commit will change that.

* Add support to settings for adding and removing controls after creation

* Rename settings with support for converting old keys

Some of these renames also come with a functional flip. This should
make things more consistently have a setting to enable a functionality
rather than to disable it.

* Manage chrome offset bugfix with settings

* Manage lookup hide sensitive filter with settings

Along with this comes the ability to automatically remove controls that
are removed from the DOM so that we don't leak them. To match this,
listeners can also now be unbound and the binding function renamed from
`change` to `listen`.

* Manage chat settings using new settings code

* Fix default settings for theme and cursor/reticule

* Add minor JSDoc to settings and rename SettingType to be clearer

* Add developer documentation for settings

* Move `webkitBased` and `possibleMobile` into semi-global scope

* Remove `disconnectOnRemove` functionality when adding a setting control

While this functionality did make it easier for calling code, it has
significant performance implications. Calling code is now expected to
manually call `remove` for controls when they are removed from the DOM.

* Use ls.has when checking for existing settings keys

* Fix canvas lock being toggled after update to new settings

This occured because `canvas.unlocked` was mistakenly not included in
the flipped mappings list when converting old settings.

* Remove admin setting for "Disable hand reset"

This setting has been redundant with the one existing in the general
settings menu for a long time now.

* Listen for chat settings close with modal events

Previously MutationObserver was used in favour of doing this to more
reliably catch when the controls should be cleaned up. Upon debate, it
was deemed that listening for the modal close event was a cleaner way to
do this and the risk posed is minimal.