Improve semantic structure of settings inputs (#368)
* Improve semantic structure of settings inputs

This does several things to improve the semantics and structure of
input fields used for settings as well as improving consistency.

- Input elements are now wrapped in a label.
- Labels containing an input element have their text content wrapped
  by an element with the `label-text` class.
- Setting fields are now grouped using paragraph elements.
- Extra text that provide additional information about a field (such
  as warnings and error messages) is now a label for the input of that
  field.

In addition to these changes, a couple of specific changes were made
to individual setting fields:
- The bubble position setting is now a simple select input.
- The template position label was broken into horizontal and vertical
  parts.

Finally, `button-bar`s were changed from being generic to using lists.
This change was incited by requiring an element allowed inside of
paragraph elements be used. `ul` worked for this and also holds some
useful semantic meaning given that `button-bar`s are flat, unordered,
sets of usually two or more buttons.

* Remove option-bubble styling

Since this is no longer a special-case component, the specific styling
for it can be removed.

* Use `div` instead of `p` for settings groups

After a lengthy and pleasant discussion about deep philosphical
matters - such as: what even *is* a paragraph with little-or-no direct
text content? What defines a list? When are semantic elements actually
important? How *would* a blind person use this art-based site? - an
agreement was made that paragraph elements suck because they can't
contain any block content. This includes lists, meaning that
`button-bar` would not work inside of a setting group without some
semantically inline element - which would be bad, semantically
speaking.

With this revelation, it was decided that simple `div`s will do here.

* Revert `button-bar` to using `div`

While I maintain that button bars are lists semnatically, the extra
level of complexity and nesting using lists here entails is probably
more un-desired than not being as semantic as possible.

* Use destructuring for generating checkbox groups