Make brightness-fixer fill board screen
Since previously this element only had the default dimensions, it could
be culled early by the browser and ignored when rendering. This resulted
in it not serving the intended purpose of subtly tricking the render
engine into working properly on Chrome.

This change uses CSS to make the element always match the size of the
canvas, ensuring that it is always on screen when needed.