Implement simple downscaling algorithm
This performs a linear average on all channels for each sampled pixel,
discarding transparent samples. This should work very nicely for
templates that only use 1-bit alpha.

A notable problem with assuming this will be the case is that the
popular pxlsfiddle tool generates templates with a semi-transparent
pixel in the top-leftmost pixel to store metadata.