Index template style by palette index
This requires storing the palette index between the template render
passes. In order to keep the color values, the original template color
values are stored in the color channels of output in the first pass.
This leaves the alpha channel open for use as the palette index. Since
pxls is currently limited to a palette size of 255, storing it in a
single color channel is possible.

One limitation with this is that storing pixels that are not on the
palette or pixels that are fully transparent introduces two new indicies
to keep track of. Transparent pixels are not so much of a problem - pxls
already uses 255 to represent these pixels, replacing an index in the
palette. This makes it pretty safe to do so here as well.

The simplest solution to storing an off-palette index is to do as with
the transparent pixels and replace one of the highest indicies. Rather
than choose 254 for this, I decided it would make more sense to use 255
which puts the style for these pixels in the very bottom-right of the
index atlas.

When thinking about where to put things in this atlas, it seemed to make
more sense to me to have the last index be used for something which is
actually used (the transparent index, 254 will never display even if
there is a pattern there). The atlas itself is a 16 by 16 grid of
styles. Palette indcices are adressed left-to-right, top-to-bottom,
putting palette index 0 at the top-left, palette index 15 at the top
right, palette index 16 one-down on the left, and so-on.

Styles currently control only opactity and so images used will have all
channels ignored besides the alpha channel. The current method of
specification is with an HTMLImageElement which is expected to be loaded
already when passed as a style (note that images with 'data:' sources
are not immediately loaded).

This specification method is convenient in some ways (allowing
web-images) and inconvenient in others (requiring async waiting,
cross-origin issues, potential difficulty of modification). It should be
noted that this format is by no-means final and other formats - such as
ArrayBuffers - should be considered.