Account for pxlsfiddle scale pixel when downscaling templates
In order to automatically detect if an image was previously processed at
a custom scale, the popular third-party tool pxlsfiddle embeds the scale
of images it processes into the top-left pixel of template images it
outputs. The embedding first checks the alpha of the existing pixel
there and then either adds or subtracts the scale based on if the
existing alpha is less than 255 - 13 or not.

13 is the arbitrary value chosen to be the maximum template scale by
pxlsfiddle, but this was previously 11 so it might be possible that
there are templates out there which seem like scale factor 1 or 2 when
they should be 12 and 13 respectively.

Regardless, this means that the simplest way to determine if this pixel
is a special scale-indicating one is to check for the alpha being
abnormal. In other words, ignore the first subpixel of a template if the
alpha is not either exactly 1 or 0.