Add simple palette conversion for templates
This adds the ability to enable color conversion as well as a simple
custom algorthm for calculating the difference between two colors. In an
ideal world, this would probably be something like CIEDE2000 but
implementing this proved to be too error-prone for me to have confidence
in implementing it correctly in GLSL.

The way in which this conversion has been added leaves plenty of option
for other color difference functions in future should someone be willing
to implement a more accurate one.