Support using CORS-proxys for template images
This adds two new config options under the new "cors" section:
- proxyBase
- proxyParam
Both of these are strings.

The "proxyBase" option takes a URL which points to a CORS proxy. The
default in the reference config is a popular public cors proxy, but
instances might choose to use their own server as a proxy if they want.
Configuration of this is left to the owner of such an instace and no
proxy server is actually implemented with this commit.

The "proxyParam" option allows the specification of an HTTP parameter
which will be set to the URI of the requested resource. If this option
is the empty string - as it is by default - the URI will be joined to
the "proxyBase" option with a forward slash before being requested.