Ensure serviceWorkerHelper.isInit is set before we start using the service worker. (#460)
Fixes a bug with multi-tabs where uiHelper.tabId would never be set because the client was preventing itself from sending the 'request-id' message to the service worker. This was caused due to a race condition in the fullfilment of the service worker registration promise (`navigator.serviceWorker.register(...).then(...)`) and the fullfilment of the ready promise (`navigator.serviceWorker.ready.then(...)`).
With this commit, `serviceWorkerHelper.isInit` is set automatically if `navigator.serviceWorker.controller` is set (if the service worker is in the active state) or, as a fail safe, on our wrapper for `navigator.serviceWorker.ready`.