Send non-redacted chat messages to staff members when snip mode is enabled. The messages are redacted on the client side to allow staff members to enjoy snip mode; but also keeping vital information like usernames for them to see when clicking on "-snip-" usernames.
This also shifts the way recognizable information in snip mode is redacted out of the objects containing the information itself.
From now on these objects contain an `object.asSnipRedacted()` method which recreates the object with some values redacted. The logic outside of that object's constructor must check `App.getSnipMode()` and call `object.asSnipRedacted()` when necessary.