Add support for shadowbanned messages
This commit adds a not null boolean `shadow_banned` column to the
`chat_messages` table to determine whether or not the user who sent the
message was shadow-banned at the time.

If a message is retrieved that has the `shadow_banned` column set to true and
the retrieving user also has the new `chat.history.shadowbanned` permission,
they will be able to see shadow-banned message(s) in slightly smaller
and red text. If the retrieving user does not have that permission, they
will not see any shadow-banned messages.

One side-effect that was not addressed is that if the shadow-banned user
reloads the page or opens the chat on a new browser page, their messages
will not appear. This is like how pixel shadow-bans are currently handled.
It would be best to address both issues in a separate commit.