Build a tentative translation framework
Upsides: static pages might become more readable to non-english
speakers.
Downsides: HTML code is pretty unreadable right now.

This doesn't cover strings served by the server directly or generated by
javascript in the client.

This particular implementation is perhaps a bit questionable. To
illustrate this, the process for getting translated strings to the
client is roughly as follows:
1. `generate-potfile.js` scans template files for pebble calls to the
   `i18n` function. It stores these as a po template file.
2. Translators use this template file to create regular po files with
   translated strings.
3. `compile-localizations.js` uses `msgcat` to convert these po files to
   Java's preferred properties files. These are compiled into the jar
   resources.
4. Pebble builds templates and uses the `i18n` function to replace
   strings as it does so.
5. The client gets the translated webpage.

It would be possible to skip step 3 here by using java's native
properties format. The cost is that GNU's gettext format generally has
more support and subjectively is easier for translators to work with.

Java supports compiling translations directly as classes instead of
doing comparatively resources lookups for every string. As with
properties files, po files can be converted directly to compiled
classes or source files with the correct information using `msgfmt`.
At this point, you're hopefully asking why we aren't doing that here.
The reason is that it seems Pebble replaces the default resolver for
getting translated strings from a Java ResourceBundle. In the process, I
can only assume it breaks the compiled class variation because I
couldn't get it to read locales other than the default with this.

The javascript programs added by this are small conversion scripts.
They are written for node which is already supported for the build
process and pull in two extra dependencies to the project. The first is
a simple node package for reading and writing po files. The second is
the GNU gettext `msgfmt` program which is invoked as a sub-process.
While I haven't put either of these programs into the automated build
process, `compile-localizations.js` should probably be put there
eventually.