Prevent repeated key events from toggling overlays
Since neither JQuery or JS provides any simple way to detect if a
keypress is repeated, the solution to catching these events is to only
allow the first one and wait until a keyup signal on the same key before
allowing the next one.

In a hotkeys rework, it may be desirabe to store the key state as a map
internally, but — since the current code is so disparate — I've used two
booleans here.