Allow placement only once pixel count has updated
This should fix the longstanding bug where the count would be one higher
than it should. This is probably one of the simpler ways to do this, but
reworking the entirety of the timer is probably a better solution.

My reasoning for this fixing the bug is that the bug occurs when:
- The current time is after the locally stored cooldown.
- The client timeout for updating the final timer tick is pending.
- The server responds to a placement before the timeout happens.

By waiting for the UI update rather than the cooldown in the first of
these assumptions, the bug should dissapear 🤞