Browserify js source files in gulp
This excludes `resources/public/include/` since they need not be
compiled separately. There are no files there yet, but this should save
some compile time when there are.

Notably, this does not mean that those files are excluded from being
output. The way the current build system seems to work is:
- Maven copies `resources/` to `target/resources/`
- Gulp generates various artifacts and replaces those in
  `target/resources/`

This means that even files unprocessed by gulp end up in the output —
such as the contents of `resouces/public/include/`. Something to
investiagte later, perhaps.