Move everything into separate modules
You may notice that this is a drastic step-up from the previous commits.
This is because the previous new modules had a nice dependency tree.
The rest of the modules (the ones added in this commit) contain a lot of
circular dependencies.

Previously, circular dependencies worked thanks to JavaScript's scoping.
Since we can no longer take advantage of that, some dependencies are
delayed until the module's `init` function is called.

Ideally, there would be no such circular dependencies but that sort of
re-architechting is non-trivial and out of scope for this.