Ignore backup files
It is (or was) a common practice to backup files before modifying
them in place using command line tools. These backups would be
automatically created with the same path as the original file plus a
`~`.

`gettext` is one such family of command line tools. It is best not to
upload these backups to the repo, so here they are ignored.