Pass all user-facing strings to the function `__`
This function just returns the input with no changes. This function is a
marker for translatable strings and may later fetch string translations.

The common name for this function is usually just `_` (a single
underscore) but since a common practice is to use this for the name of
an unused variable, I felt using two underscores was a bit safer.

Another possible name for this function is `gettext` but this is rather
verbose for what is needed here, given the frequency of this functions
appearence.