Extract script strings
This could probably have been done with a regex approach as with pebble
templates, but — since I have a JS parser — it is done using esprima.
This is probably more robust while also more likely to be incorrect. It
seems to work fine but I didn't check it all that hard.

Another option to this entirely is to use an established library for
gettext, such as ttag. That would offer a number of benefits including a
plan to actually load translations back into scripts that this currently
doesn't offer. We will see, I suppose…