Compile translated script files on build
This all seems to work suprisingly well. The most worrying thing
currently is that this code makes a lot of assumptions about the
language codes used. Notably, it basically pretends that subtypes don't
exist and expects to be able to check for just the first part.

That's not the only thing worrying about all this, but what about pxls
codebase isn't?