Use pac4j for authentication
Nothing significant has changed here from a functionality standpoint,
with the exception that users can no longer sign up and logins from one
tab will no proliferate to others.

This puts the code in a good place to start making most of the huge
breaking changes though, so don't expect that to stay the same for long.