Automatically sign users up based on login
Taking a step back from breaking everything for a second since I'm
really not sure which data should go where.

This comes with a couple of database changes: the `user_logins` table is
dropped and a new field is added to the `users` table. This new field is
`sub VARCHAR(60) UNIQUE NOT NULL`. Astute readers might notice that this
structure is basically a reversion back to how things were before
multiple logins were added. This is because there is now one source of
logins — the dedicated ID server.

There is one issue with this structure — IP logins need a `sub` value
but obviously don't have one. As with the username, this is just set to
the MD5 hash of the IP. This is probably not a foolproof solution and
there could theoretically be collisions with this and the ID server's
subject identifiers, but I think I'll leave fixing that for later since
IP logins are going to be a bit hackey either way without some proper
fixing.