Sync username with ID server
This also removes all of the infrastructure for forcing user's to change
their name. The reasoning for this is that it can now be enforced by the
ID server, but there is a catch — the current session continues to be
valid until;
- a new request is made *and*
- the current access token expires.

Because pxls uses the websocket for things like placements, that means
that users can continue using the site without changing their username
indefinitely if they stay connected. I'll look into mitigations of this
soon, because I very much doubt this sort of thing is desirable.