Store template style preferences in settings
As noted in the code, the settings keys used here are a bit ugly since
there's essentially two used to store one value. This is not quite true
because of the rare case where a user has a custom template url in the
custom section but has another style selected. I don't think this an
entirely unexpected case and supporting that requires storing them
separately.