Add userkeys2db script
This script will take a user keys file and store its contents in the
database. It is very much based on the db2pixellogs script, except it
sort-of does the opposite. This includes a dry-run option though I'm not
sure how helpful that currenly is since the only output is from errors
and that rolls back the transaction anyway 🤷.