Add settings for when to change the chat icon
This totally breaks a lot of semantics in the code since now the classes
for notification and ping are disjoint from notifications and pings, but
hopefully people stop asking for a button to outright remove chat.

This is probably one of those changes that will be looked on (at least
by me) as overall detrminental rather than positive but I don't really
care at this point.