Update discord details from connected user accounts
This will always show a user's discord account if one is linked. Some
users may not desire this so work might not be done here but at least it
will show the *correct* account rather than just relying on the user to
tell the truth.

Most of the work in this went into trying to get keycloak to put
connected user accounts in the token. While technically any identity
server can work with this, it would be misleading if I didn't say it's
pretty keycloak-oriented now.

Anyway, for this to work the ID provider should offer a scope
"connected_accounts". This scope should add an `accounts` field to the
user's tokens containing an array of objects with the following entries:
- `userid`: the user's service ID
- `userName`: the user's service username
- `identityProvider`: the name of the service

In keycloak, this can be achieved with a script mapper for the custom
scope. This is what took the most time in this commit:

```javascript
providers = realm.getIdentityProviders().toArray();
ArrayList = Java.type("java.util.ArrayList");
userProviders = new ArrayList();
for(i = 0; i< providers.length; i++) {
    alias = providers[i].alias;
    provider = keycloakSession
        .users()
        .getFederatedIdentity(realm, user, alias);
    if(provider != null) {
        userProviders.add(provider);
    }
}
userProviders.toArray()
```

I know that looks wierd: keycloak's javascript engine is very brittle.

You're on your own with other ID providers — I wasn't even sure it was
possible in keycloak for a long time.