Store user links in database
This removes the `discord_name` field from the `users` table and adds a
new table: `user_links`. Discord details are still the only exposed
details, but the backend is there for supporting whatever user details
are needed.

This currently updates all the user links every time they request a
page. This is obviously too much and should be reduced to just when
their login token changes.