Make opacity a preference rather than a URL parameter (#537)
This also sets the default opacity to 1.0 since it seems that's what
most people are using anyway.

Basically all templates are distributed with a style of 1.0 and anyone
who wants to use a different opacity has to do so for every new
template. By making opacity a preference instead, template distributers
no longer need to worry about this and template consumers no longer need
to change the option to their liking for every template opened.

Additionally, this includes a fix an issue where template conversion
mode was not copied with the copy template hotkey.