Add language override setting
This allows users to tell pxls to ignore their browsers language
preference and use a specific localisation.

As much as it would be better if browsers solved this via
well-established standards, the web is a complete mess of doing things
the wrong way and that's unlikely to change anytime soon.