Add user keys for accessing sanitized data (#540)
* Update log sanitization script

Apparently IPs are no longer put into logs. This is a rather simple fix:
simply don't try and parse IPs, reducing the expected field count from 7
to 6.

* Add userkeys2db script

This script will take a user keys file and store its contents in the
database. It is very much based on the db2pixellogs script, except it
sort-of does the opposite. This includes a dry-run option though I'm not
sure how helpful that currenly is since the only output is from errors
and that rolls back the transaction anyway 🤷.

* Add user log keys

This adds a new database table as well as a new tab in the profile
page for keys used in pixel logs.

* Add user ID to log handling scripts

Relying on username alone is prone to more failure than the more
reliable user id. For any humans reading the logs, usernames are kept
for convenience.

* Use key icon for profile data tab

Co-authored-by: Vanilla <vanilla@f66.dev>