Remove css vars from complex position calculations
Apparently some browsers choke on this, resulting in broken panels. This
makes the code a bit harder to maintain, but shouldn't break anything
functionally.

Notably, this probably doesn't completely solve all issues with the
faulty browser, since it also seems to dislike the min/max calculations.
That said, it seemed to only choose the wrong path for those rather than
refusing to work like with variables, so this should still work in most
cases 🤞.