Corrections and additions to the Russian translation
There are quite a few changes

Recently brought a couple of Russian-speaking friends who had to be literally dragged through the settings.

I myself use the English version of the site because the Russian localization is almost unedited for users and very incomprehensible. 

I made it easier to understand and closed some empty moments.