Avoid querying overlay data too many times
Previously because of the creation of 2 virginmap overlays and 2
heatmap overlays, /virginmap and /heatmap would be each called twice,
and /placemap would be called 4 times.
What we can do instead is to call each of these endpoints once, and
reuse the relevant responses across all overlays.

Due how much data is needed for each of these endpoint calls (currently
1 byte per pixel in the canvas), the elimination of 5 useless calls
will result in significant data usage reduction, and possibly a speedup
in the canvas rendering.