Really initialize overlays lazily
The previous lazy initialization would always be called at canvas
initialization regardless of whether the overlays should be shown or
not.
This adds a check such that overlays do not initialize until they need
to be shown.

This should save a lot on data usage when overlays are not used (which
tends to be the case on mobile when data is limited).