Deduplicate overlay data fetching
When displaying virgin map and heat map, we do not need to fetch
/placemap 4 times, /virginmap 2 times and /heatmap 2 times.
We can instead fetch each of these once, reusing the relevant responses
across overlays.

This also adds a utility class LazyPromise to allow creating the fetch
promises without starting their requests until it is needed (when the
promise is awaited).