Fix error when displaying a truncated template url in chat
When a template url is truncated, it can cause its name to become an
invalid URI component if a percent encoded character is truncated in
the middle.
This causes the whole message to fail the markdown processing and be
received as a raw string.
We now catch that error and fall back to the (improperly) encoded name,
which may look weird but is still better than the raw url.