Update pac4j and generally fix and improve OpenID handling
This includes a change to how the `host` config value is interpreted. It
should now be a URL to the client base URL. This helps in pretty much
all scenarios where the parameter is used but will require migration.

Since it was necessary for testing, this also comes with a hidden new
config option `auth.devmode` which will disable ssl verification and
print warnings related to invalid openid profiles.