Revert deletion of rename code
Trying to keep the username in sync with the preferred username from the
ID server is a rather poor way to handle username and other user data.
In light of that, the existing mechanisms for handling usernames should
no longer be removed by this branch.

Since usernames are no longer necessarily unique, some of the handling
needs to change to accomodate this. It's possible uniqueness may
re-added but I'd like to handle that seperately.